+++ src/main/java/com/drmangotea/tfmg/content/machinery/vat/base/VatBlockEntity.java
@@ -5,6 +5,7 @@ import com.drmangotea.tfmg.base.TFMGBlockConnectivityHandler;
 import com.drmangotea.tfmg.base.TFMGUtils;
 import com.drmangotea.tfmg.base.capabilities.TFMGCapabilities;
 import com.drmangotea.tfmg.base.capabilities.pressure.IPressureHandler;
+import com.drmangotea.tfmg.base.lang.TFMGLang;
 import com.drmangotea.tfmg.base.lang.TFMGTexts;
 import com.drmangotea.tfmg.base.pressure.Pressure;
 import com.drmangotea.tfmg.base.pressure.behaviour.SmartPressureTankBehaviour;
@@ -75,7 +76,6 @@ import java.util.*;
 import static java.lang.Math.abs;
 
 public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInformation, IMultiBlockEntityContainer.Fluid, Clearable {
-
     private static final int MAX_SIZE = 3;
     private static final int MAX_PRESSURE = 30;
 
@@ -101,6 +101,7 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
     protected int width;
     protected int height;
     //updating and technical stuff
+    protected int recipeDuration;
     protected BlockPos controller;
     protected BlockPos lastKnownPos;
     protected boolean updateConnectivity;
@@ -205,6 +206,12 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
         TFMGBlockConnectivityHandler.formMulti(this);
     }
 
+    public int getProgressPercentage() {
+        if (recipeDuration <= 0)
+            return -1;
+        return Math.min(100, (int) (100f * timer / recipeDuration));
+    }
+
     //goggle stuff
     public MutableComponent getHeatComponent(boolean forGoggles) {
         return componentHelper("heat", heatLevel, forGoggles);
@@ -212,6 +219,13 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
 
     public MutableComponent getPressureComponent(boolean forGoggles) {
         return componentHelper("pressure", pressure.getValue(), forGoggles);
+    }
+
+    public MutableComponent getProgressComponent() {
+        int progress = getProgressPercentage();
+        if (progress == -1)
+            return null;
+        return TFMGLang.translateDirect("goggles.progress", Component.literal(getProgressPercentage() + "%").withStyle(ChatFormatting.GOLD)).withStyle(ChatFormatting.GRAY);
     }
 
     private MutableComponent componentHelper(String label, int level, boolean forGoggles, ChatFormatting... styles) {
@@ -917,7 +931,7 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
                 }
             }
         }
-        if (oldMachineMap != machineMap)
+        if (!oldMachineMap.equals(machineMap))
             recipe = null;
 
         notifyUpdate();
@@ -1216,8 +1230,12 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
         if (!isController())
             return getControllerBE().addToGoggleTooltip(tooltip, isPlayerSneaking);
 
-        TFMGTexts.header("vat").style(ChatFormatting.GRAY)
-                .forGoggles(tooltip);
+        TFMGTexts.header("vat").style(ChatFormatting.GRAY).forGoggles(tooltip);
+
+        MutableComponent progressComp = getProgressComponent();
+        if (progressComp != null) {
+            CreateLang.builder().add(getProgressComponent()).forGoggles(tooltip, 1);
+        }
 
         CreateLang.builder().add(getPressureComponent(true)).forGoggles(tooltip, 1);
         CreateLang.builder().add(getHeatComponent(true)).forGoggles(tooltip, 1);
@@ -1276,6 +1294,7 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
             outputInventory.deserializeNBT(registries, compound.getCompound("OutputItems"));
             timer = compound.getInt("Timer");
             heatLevel = compound.getInt("HeatLevel");
+            recipeDuration = compound.getInt("RecipeDuration");
             pressure = Pressure.from(compound);
         }
 
@@ -1327,6 +1346,7 @@ public class VatBlockEntity extends SmartBlockEntity implements IHaveGoggleInfor
             compound.put("OutputItems", outputInventory.serializeNBT(registries));
             compound.putInt("Timer", timer);
             compound.putInt("HeatLevel", heatLevel);
+            compound.putInt("RecipeDuration", recipe != null ? recipe.getProcessingDuration() : 0);
             pressure.save(compound);
         }
         compound.putInt("Luminosity", luminosity);